requirements.txt
# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Data & Image Processing
Pillow>=9.0.0
numpy>=1.24.0
scipy>=1.10.0
scikit-image>=0.20.0

# Evaluation Metrics
clean-fid>=0.1.35
lpips>=0.1.4
pytorch-fid>=0.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
imageio>=2.31.0
imageio-ffmpeg>=0.4.8

# Configuration & Utilities
PyYAML>=6.0
tqdm>=4.65.0
tensorboard>=2.13.0

# Kaggle API
kaggle>=1.6.0

# Testing
pytest>=7.3.0
